+++ extensions/jose/CHANGELOG.md
@@ -5,3 +5,8 @@
 - Assistant 
 - Snippets
 - Chat conversation
+
+## [Version: 1.0.1] - 2024-07-04
+
+- update default assistant
+- add default snippet